django_service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
cryptography==42.0.4
PyMySQL==1.1.0
pika==1.3.2
drf-spectacular==0.27.1
python-dotenv==1.0.1
gunicorn==20.1.0
